feat: AppRestorationService takes care of respring the app
